Replaced singularity with apptainer in the exercise
The text says 

```The tutorial container is based on Ubuntu 18.04. The host and the container use the same kernel, but the rest of the system can vary.```

but puhti doesn't run Ubuntu. I believe this should be modified to 

```
The host uses Red Hat 8 with the Linux kernel 4.18. 
```

The kernel version could actually be omitted. New users likely have no idea what it means.